Add two helper functions split and strip for future refactoring
First function (split) accepts a string and a character and splits string on that specific
character to a vector of strings and return the vector.

Second function (strip) accepts a string and return a copy of the string in which all whitespaces
have been stripped from the beginning and the end of the string.

Implement set_imj method in Complex class which was deleted by mistake in a previous commit.